refactor(frontend): split six oversized files below the 400-line structure threshold
