+++ apps/desktop/src/components/settings/panes/OrganizationPane.tsx
@@ -10,14 +10,14 @@ import {
   useGitHubAppInstallationStatus,
   useStartGitHubAppInstallation,
 } from "@proliferate/cloud-sdk-react";
-import { ProviderBrandIcon } from "@proliferate/product-ui/auth/ProviderBrandIcon";
-import { Badge } from "@proliferate/ui/primitives/Badge";
 import { UpgradeGateDialog } from "@/components/billing/UpgradeGateDialog";
 import { OrganizationBillingLinkSection } from "@/components/settings/panes/organization/OrganizationBillingLinkSection";
 import { OrganizationSettingsCard } from "@/components/settings/panes/organization/OrganizationSettingsCard";
 import { OrganizationSection } from "@/components/settings/panes/organization/OrganizationLogo";
-import { SettingsSection } from "@proliferate/product-ui/settings/SettingsSection";
-import { SettingsRow } from "@proliferate/product-ui/settings/SettingsRow";
+import {
+  GitHubAppInstallationSection,
+  isOrganizationAdminRole,
+} from "@/components/settings/panes/organization/GitHubAppInstallationSection";
 import { SettingsPageHeader } from "@proliferate/product-ui/settings/SettingsPageHeader";
 import { SettingsEmptyState } from "@proliferate/product-ui/settings/SettingsEmptyState";
 import { useOrganizationActions } from "@/hooks/access/cloud/organizations/use-organization-actions";
@@ -314,84 +314,6 @@ export function OrganizationPane() {
       ) : null}
     </section>
   );
-}
-
-function GitHubAppInstallationSection({
-  loading,
-  installing,
-  canManage,
-  status,
-  onInstall,
-  onManage,
-}: {
-  loading: boolean;
-  installing: boolean;
-  canManage: boolean;
-  status: {
-    installed: boolean;
-    accountLogin?: string | null;
-    repositorySelection?: string | null;
-    suspendedAt?: string | null;
-  } | undefined;
-  onInstall: () => void | Promise<void>;
-  onManage: () => void | Promise<void>;
-}) {
-  const installed = status?.installed === true;
-  const repositorySelection = formatRepositorySelection(status?.repositorySelection ?? null);
-  const statusLabel = loading
-    ? "Checking"
-    : installed
-      ? "Installed"
-      : "Not installed";
-  const description = installed
-    ? `Installed on ${status?.accountLogin ? `@${status.accountLogin}` : "this GitHub account"}${repositorySelection ? ` with ${repositorySelection}` : ""}.`
-    : canManage
-      ? "Install the Proliferate GitHub App for this organization before members can enable cloud repositories."
-      : "Ask an organization admin to install the Proliferate GitHub App before you enable cloud repositories.";
-
-  return (
-    <SettingsSection
-      title="GitHub App"
-      description="Repository access for organization cloud environments."
-    >
-      <SettingsRow
-        label="Organization installation"
-        description={description}
-      >
-        <div className="flex flex-wrap items-center justify-end gap-2">
-          <Badge tone={installed ? "success" : "warning"}>{statusLabel}</Badge>
-          {canManage ? (
-            <Button
-              type="button"
-              variant="secondary"
-              loading={installing}
-              disabled={installing}
-              onClick={() => {
-                void (installed ? onManage() : onInstall());
-              }}
-            >
-              {!installing ? <ProviderBrandIcon provider="github" className="size-[13px]" /> : null}
-              {installed ? "Manage in GitHub" : "Install GitHub App"}
-            </Button>
-          ) : null}
-        </div>
-      </SettingsRow>
-    </SettingsSection>
-  );
-}
-
-function isOrganizationAdminRole(role: string | null | undefined): boolean {
-  return role === "owner" || role === "admin";
-}
-
-function formatRepositorySelection(repositorySelection: string | null): string | null {
-  if (repositorySelection === "all") {
-    return "all repositories";
-  }
-  if (repositorySelection === "selected") {
-    return "selected repositories";
-  }
-  return null;
 }
 
 function OrganizationNotice({ children }: { children: ReactNode }) {
